Move product fixture generator
