+++ corehq/apps/products/fixtures.py
@@ -1,0 +1,17 @@
+from corehq.apps.products.models import Product
+from corehq.apps.commtrack.fixtures import _simple_fixture_generator
+
+
+def product_fixture_generator(user, version, synclog, last_sync):
+    fields = [
+        'name',
+        'unit',
+        'code',
+        'description',
+        'category',
+        'program_id',
+        'cost',
+        'product_data'
+    ]
+    data_fn = lambda: Product.by_domain(user.domain, include_archived=True)
+    return _simple_fixture_generator(user, "product", fields, data_fn, last_sync)
